🐛 Update of PROPFIND method and connecting it to Documents service
