+++ tdrive/backend/node/src/core/platform/framework/logger.ts
@@ -12,12 +12,6 @@ export const logger = pino({
   mixin() {
     return executionStorage.getStore() ? executionStorage.getStore() : {};
   },
-  transport: {
-    target: 'pino-pretty',
-    options: {
-      colorize: true
-    }
-  },
 });
 
 export const getLogger = (name?: string): TdriveLogger =>
